refactor(tests): split engine tests to run in parallel worker-farm
